修改机器人基本设置是会更新到solr中
git-svn-id: https://svn.chsi.com.cn/svn/repos/chsicommunity/knowledge@23756 5607bd42-195c-0410-82fb-b7440d194b30

+++ trunk/webapp/src/main/java/com/chsi/knowledge/admin/action/RobotSetAction.java
@@ -181,9 +181,11 @@ public class RobotSetAction extends AjaxAction{
             robotIndexData.setA(anser);
             robotIndexData.setNum(num);
             robotService.deleteRobotQASet(id);
-            SearchServiceClient searchClient = SearchServiceClientFactory.getSearchServiceClient();
-            searchClient.updateRobotQA(robotIndexData);
-            searchClient.deleteRobotQA(id);
+            if(!robotIndexData.getQ().startsWith("#")){
+                SearchServiceClient searchClient = SearchServiceClientFactory.getSearchServiceClient();
+                searchClient.updateRobotQA(robotIndexData);
+                searchClient.deleteRobotQA(id);
+            }
             ManageCacheUtil.removeRobotABySpecialQ(q);
         }
         return SUCCESS;
